[Refactor] ArticleResponse를 ArticleQuery 내부에 구현
사용처가 한정적인 dto를 dao 내부에 구현하므로써 관리용이성 추구
