+++ src/main/java/codesquad/servlet/QnaServlet.java
@@ -6,7 +6,7 @@ import codesquad.domain.user.User;
 import codesquad.exception.UnauthorizedRequestException;
 import codesquad.servlet.annotation.authentication.Authorized;
 import codesquad.servlet.dao.ArticleQuery;
-import codesquad.servlet.dto.ArticleResponse;
+import codesquad.servlet.dao.ArticleQuery.ArticleResponse;
 import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletException;
