+++ src/main/java/codesquad/servlet/dao/ArticleQuery.java
@@ -1,7 +1,5 @@
 package codesquad.servlet.dao;
 
-import codesquad.servlet.dto.ArticleResponse;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -9,4 +7,13 @@ public interface ArticleQuery {
     Optional<ArticleResponse> findById(Long id);
 
     List<ArticleResponse> findAll();
+
+    record ArticleResponse(
+            Long articleId,
+            String title,
+            String content,
+            Long writerId,
+            String writer
+    ) {
+    }
 }
