--- src/main/java/codesquad/servlet/dto/ArticleResponse.java
@@ -1,9 +1,0 @@
-package codesquad.servlet.dto;
-
-public record ArticleResponse(Long articleId,
-                              String title,
-                              String content,
-                              Long writerId,
-                              String writer
-) {
-}
