+++ src/main/webapp/WEB-INF/views/qna/update-form.jsp
@@ -1,4 +1,4 @@
-<%@ page import="codesquad.servlet.dto.ArticleResponse" %>
+<%@ page import="codesquad.servlet.dao.ArticleQuery.ArticleResponse" %>
 <%@ page contentType="text/html; charset=UTF-8" %>
 <!DOCTYPE html>
 <html lang="kr">
