+++ src/test/java/codesquad/servlet/MockArticleQuery.java
@@ -1,7 +1,6 @@
 package codesquad.servlet;
 
 import codesquad.servlet.dao.ArticleQuery;
-import codesquad.servlet.dto.ArticleResponse;
 
 import java.util.Collections;
 import java.util.List;
